refactor(extensions): move AI provider + embedding + registry into src/extensions/ai/
- New src/extensions/ai/{ai-provider,embedding-provider,ai-provider-registry,index}.ts
- Barrel uses `export type {}` for interfaces (the spec called for plain
  `export {}`, but under Deno `--no-check` SWC transpiles files in isolation
  and emits a runtime value re-export for interface names, which fails ESM
  resolution; `export type {}` is the working pattern). Plain `export {}`
  reserved for runtime values (AIProviderRegistryName, createAIProviderRegistry).
- Legacy src/extensions/interfaces/index.ts kept as back-compat shim
  pointing at ../ai/index.ts.
- src/extensions/builtin-extensions.ts + src/server/bootstrap.ts now
  import AI symbols from the new ai/ barrel.
- tests/_helpers/context.ts + src/extensions/{integration,orchestrate}.test.ts
  updated to the new path.
- deno.json adds ./extensions/ai export.
