+++ deno.json
@@ -69,6 +69,7 @@
     "./embedding": "./src/embedding/index.ts",
     "./extensions": "./src/extensions/index.ts",
     "./extensions/interfaces": "./src/extensions/interfaces/index.ts",
+    "./extensions/ai": "./src/extensions/ai/index.ts",
     "./testing": "./src/testing/index.ts",
     "./testing/assert": "./src/testing/assert.ts",
     "./testing/bdd": "./src/testing/bdd.ts",
